move useXTerm to hooks folder
