+++ frontend/src/hooks/useXTerm.test.ts
